Large Update
Uploading my current work on the recode for GrimList.

Several new ways of going about coding can be seen in this update. No
real functionality, as of yet.

MetricsManager and UpdateManager should be complete.
Added placeholders for all of the possible commands.

Add / Remove commands are... functional, but aren't useful as they don't
actually whitelist anyone, yet. But the logic behind it should be final.