UpdateManager / Updater commit
Another update to the recoding being done to GrimList.

I lied about UpdateManager and MetricsManager being finished. Although
the change done to MetricsManager was minimal. Just a variable
name-change.

UpdateManager had an additional feature added that would warn the
console of an update, once the server starts. Additionally, a few unused
variables and grabs from Updater were removed.

PlayerManager had a tiny update done as well, which prettifies the
/whitelist update command.

Executor notices a tiny change to what information gets sent to through
to the Update command, in preparation for allowing additional input for
the /whitelist update command.