Large Updates (and some tiny ones)
Update to ConfigManager. Decided to use it as the SQL Manager. Cuts down
on code a bit, and leaves out needing another class file.

Unless there are bugs (which is possible, as this commit hasn't been
tested), "Add.java" is completed. I think it looks rather nice.

toExecutor() was stupid. That was removed.

TINY copyright update. Accounts for the majority of the edited files.