Resource Omitting
Realized, now, that I've been carrying Metrics/Updater.java in my
available source code here on GitHub. I never noticed, before now, and
apologize to the respective creators for my error.

Shouldn't be an issue from here on out. Each of them have been moved to
a new source folder, and omitted in .gitignore.

Overall, no impact on the rest of the plugin.