SaveQueries / GetUUID Update
This update adds functionality to save UUIDs / usernames into playerdata
when queries are made for those UUIDs (but not already saved,
afterwards, such as adding to whitelist).

This way, you can speed up future searches, without having to have that
player have been on the whitelist at some point. This is optional, and
can be turned off in the config. It is true, by default.

Additionally, there is now a 'getid' function, which will allow a user
to see search for a player's UUID by their username. This will hook into
SaveQueries, if enabled, and save that information to playerdata.

As you can see from an alias assembler in previous versions of
Executor.java, this was already a planned feature.