Quickie Update
Removing commandusage.yml. An unused resource, now that I've decided to
forgo recording command logs, for files. This may see a return, in the
future.

Updated plugin.yml with the appropriate permissions. Don't know if I did
a replace all flub, and renamed everything to whitelist, or if I just
did that on accident while not paying attention, some other day. But,
it's fixed, now.