PlayerData merged into FileManager
Since the only thing that uses PlayerData is FileManager, it made a lot
more sense to just combine them, and save the worry about constructing a
non-confusing method, and ensuring that nothing is ever null.

Now, once FileManager is called from the main class, it's good for the
rest of it's use. Plus, it shortens the code a bit in FileManager, when
attempting to write to file.