AlwaysLookup Update
This isn't a security flaw, really. Because the PlayerManager functions
by grabbing the UUID of the player, and checking against Player Data,
there's no way for this to be exploited.

However, I'm implementing AlwaysLookup, as an optional feature.

Why?

Commands could, theoretically, be applied to the wrong person. For
obvious reasons. AlwaysLookup will prevent this.