The MySQL Update
Large commit. Mostly focused on adding MySQL functionality. I'll go over
the other updates, as MySQL alone is pretty straight forward :

Firstly, UUIDFetcher was updated! Queries against the Mojang servers
seem to be much faster, now. Thanks, evilmidget38!

Secondly,  LogUsage. Currently, MySQL is the only focus which can use
this, but I'll expand it to other ones, in the future. If set to true,
it'll log every command used using this plugin, for the respective
command.