MySQL Bugfixes
So, that's all the bugs I've been able to find on my own. =/

I feel really dirty about releasing it in this state. But it works, and
I haven't been able to throw any bugs through my light testing. But I'm
hardly a stress tester. I'm sure someone will find something.

Because it's a MySQL release, and pretty massive, I'll be upping to
3.1.0 after this.