MySQL update
Removed the registerPaths() method, which was similar to the
registerPaths() method in FileManager.

It's a good idea, for sure. But, considering that registerPaths needs to
be called every time a query is run (otherwise, you risk issues when
changing configuring within the game), and most of the
ps.executeUpdate() queries only need the database as a variable, it made
much more sense to just draw the database from config each time, instead
of drawing a the entire configuration.

Also, I've removed the executeUpdate() method. Now, all queries are run
inline, instead of referring to an outside class.

executeUpdate() was mostly just for bug testing, as I found it to be
more human readable. This shouldn't be an issue, anymore, as any bugs at
this point _should_ be easy to fix.