SetConfig bugfixes
Just some silly mistakes. New methods weren't saving. Some portions of
the new methods also didn't return true, so you were getting the
successful message, as well as the error message (because it would
proceed to the end of the method, unchecked).

Should be fixed now. Can't find any more bugs, at the moment.

This release should lead up to an update, as I'm confident in the MySQL
version, as it is now. And SetConfig is where I want it to be, with the
new MySQL settings working.